+++ VersionHistory.md
@@ -9,3 +9,4 @@
   3. Version 0.2  
     - 2015-09-01: 0.2.0-pre - add XAudio2 2.7 support in DXSDK preliminarily
     - 2015-09-02: 0.2.0 - add XAudio2 2.9 support 
+    - 2015-09-03: 0.2.1 - generalized error message